refactor: refactored SeasonsScreen.
