--- ui/src/main/java/com/cairosquad/ui/details/SeasonsScreen.kt
@@ -1,163 +1,0 @@
-package com.cairosquad.ui.details
-
-import android.os.Build
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.statusBars
-import androidx.compose.foundation.layout.systemBars
-import androidx.compose.foundation.layout.windowInsetsPadding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.BlurredEdgeTreatment
-import androidx.compose.ui.draw.blur
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.cairosquad.design_system.R
-import com.cairosquad.design_system.basic_component.AppBar
-import com.cairosquad.design_system.modifier.dropShadow
-import com.cairosquad.design_system.theme.Theme
-import com.cairosquad.ui.movio_component.LoadingMovieImage
-import com.cairosquad.ui.movio_component.SeasonCard
-import com.cairosquad.ui.navigation.EpisodesRoute
-import com.cairosquad.ui.navigation.LocalNavController
-import com.cairosquad.ui.utils.ObserveAsEffect
-import com.cairosquad.viewmodel.details.series.season.SeasonDetailEffect
-import com.cairosquad.viewmodel.details.series.season.SeasonDetailsInteractionListener
-import com.cairosquad.viewmodel.details.series.season.SeasonDetailsScreenState
-import com.cairosquad.viewmodel.details.series.season.SeasonsViewModel
-import kotlinx.coroutines.Dispatchers
-
-@Composable
-fun SeasonsScreen(
-    seriesId: Long,
-) {
-    val viewModel: SeasonsViewModel =
-        hiltViewModel<SeasonsViewModel, SeasonsViewModel.Factory> { factory ->
-            factory.create(
-                seriesId = seriesId,
-                dispatcher = Dispatchers.IO
-            )
-        }
-    val uiState by viewModel.screenState.collectAsStateWithLifecycle()
-    val navController = LocalNavController.current
-
-    ObserveAsEffect(viewModel.effect) { effect ->
-        when (effect) {
-            SeasonDetailEffect.NavigateBack -> navController.popBackStack()
-            is SeasonDetailEffect.NavigateToEpisodesScreen -> navController.navigate(
-                EpisodesRoute(
-                    effect.seriesId,
-                    effect.seasonNumber
-                )
-            )
-
-        }
-    }
-
-    SeasonScreenContent(
-        uiState = uiState,
-        listener = viewModel,
-    )
-}
-
-@Composable
-fun SeasonScreenContent(
-    uiState: SeasonDetailsScreenState,
-    listener: SeasonDetailsInteractionListener,
-) {
-    Box {
-        Box(
-            modifier = Modifier
-                .windowInsetsPadding(WindowInsets.statusBars)
-                .size(230.dp)
-                .align(Alignment.TopEnd)
-                .then(
-                    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.S) {
-                        Modifier.dropShadow(
-                            shape = CircleShape,
-                            color = Theme.color.surfaces.onSurfaceAt5,
-                            blur = 264.dp,
-                            offsetX = 0.dp,
-                            offsetY = 0.dp,
-                            alpha = 0.10f
-                        )
-                    } else {
-                        Modifier
-                            .blur(264.dp, edgeTreatment = BlurredEdgeTreatment.Unbounded)
-                            .background(
-                                color = Theme.color.surfaces.onSurfaceAt5,
-                                shape = CircleShape
-                            )
-                    }
-                )
-        )
-        Column( modifier = Modifier
-            .fillMaxWidth()
-            .windowInsetsPadding(WindowInsets.systemBars)
-        ) {
-            AppBar(
-                title = stringResource(R.string.current_season),
-                onBackButtonClicked = listener::onBackClicked,
-            )
-        LazyColumn(
-            modifier = Modifier
-                .fillMaxSize(),
-            verticalArrangement = Arrangement.spacedBy(12.dp),
-            contentPadding = PaddingValues(start = 16.dp, end = 16.dp, bottom = 16.dp)
-        ) {
-            when (uiState.seasonSectionState) {
-                SeasonDetailsScreenState.ScreenStatus.LOADING -> {
-                    items(10) {
-                        LoadingMovieImage(
-                            Modifier
-                                .height(100.dp)
-                                .fillMaxWidth()
-                        )
-                    }
-                }
-
-                SeasonDetailsScreenState.ScreenStatus.SUCCESS -> {
-                    items(uiState.season) { season ->
-                        SeasonCard(
-                            seriesName = uiState.seriesTitle,
-                            seasonTitle = season.name,
-                            seasonRate = season.rating,
-                            totalNumberOfEpisodes = season.episodesCount.toString(),
-                            movieImage = season.posterPath,
-                            yearOfPublish = season.airDate,
-                            timeOfPublish = season.timeOfPublish,
-                            currentSeason = "${season.number}",
-                            onClick = { listener.onSeasonClicked(season.seriesId, season.number) },
-                            modifier = Modifier
-                                .height(100.dp)
-                                .fillMaxWidth(),
-                        )
-                    }
-
-                }
-                SeasonDetailsScreenState.ScreenStatus.ERROR -> {
-                   item {
-                       DetailsFailContent(onTryAgainClick = listener::onRefresh)
-                   }
-
-                }
-            }
-        }
-        }
-    }
-}
